Moved shared nma calls into ENM base class
